+++ src/_priv/mod.rs
@@ -1,0 +1,4 @@
+pub use std;
+pub use std::ffi::c_void;
+
+pub mod cfstring;